inpainting_evaluation/requirements.txt
accelerate==0.22.0
albumentations==1.3.1
controlnet-aux==0.0.7
datascience==0.17.6
diffusers==0.20.2
huggingface-hub==0.17.1
imageio==2.31.3
keras==2.13.1
matplotlib==3.7.1
mmagic==1.0.2
mmcv==2.0.1
mmengine==0.8.4
notebook==6.5.5
numpy==1.23.5
opencv-python==4.8.0.76
openmim==0.3.9
pandas==1.5.3
Pillow==9.4.0
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.2.2
scipy==1.11.2
tensorboard==2.13.0
tensorflow==2.13.0
torch==2.0.1
torchaudio==2.0.2
torchdata==0.6.1
torchsummary==1.5.1
torchtext==0.15.2
torchvision==0.15.2
transformers==4.33.1
